% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_graph_series}
\alias{check_graph_series}
\title{Check that an object is a social_graph_series}
\usage{
check_graph_series(x)
}
\arguments{
\item{x}{Object to check.}
}
\description{
Check that an object is a social_graph_series
}
\keyword{internal}
