% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-checks.R
\name{check_igraph}
\alias{check_igraph}
\title{Check that an object is an igraph}
\usage{
check_igraph(x)
}
\arguments{
\item{x}{Object to check.}
}
\description{
Check that an object is an igraph
}
\keyword{internal}
