% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{gini_coeff}
\alias{gini_coeff}
\title{Compute Gini coefficient}
\usage{
gini_coeff(x)
}
\arguments{
\item{x}{A non-negative numeric vector.}
}
\value{
A scalar Gini coefficient between 0 and 1.
}
\description{
Compute Gini coefficient
}
\keyword{internal}
