% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_density.R
\name{network_density_ts}
\alias{network_density_ts}
\title{Compute network density over time}
\usage{
network_density_ts(graph_series)
}
\arguments{
\item{graph_series}{A \code{social_graph_series} (output of \code{\link[=build_graph_series]{build_graph_series()}}).}
}
\value{
A tibble with columns:
\describe{
\item{period}{Time period (character, ISO date).}
\item{n_nodes}{Number of active nodes.}
\item{n_edges}{Number of edges.}
\item{density}{Network density (0--1).}
}
}
\description{
Network density is the proportion of possible edges that are present.
High density indicates a highly connected network; low density indicates
sparse interaction.
}
\examples{
data(sim_social_events)
ev <- as_social_events(sim_social_events)
gs <- build_graph_series(ev, window = "month")
network_density_ts(gs)

}
