% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{top_p_share}
\alias{top_p_share}
\title{Get top-p share of a distribution}
\usage{
top_p_share(x, p = 0.1)
}
\arguments{
\item{x}{Non-negative numeric vector.}

\item{p}{Proportion (0--1) of top actors.}
}
\value{
Share of total held by top-p proportion.
}
\description{
Get top-p share of a distribution
}
\keyword{internal}
